feat: 영상 컴포넌트 연결 #35
