+++ frontend/src/components/Monitoring/InProgress/IncidentCard.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { IoMdCamera } from 'react-icons/io';
 import FeedbackCard from './FeedbackCard';
+import VideoModal from './VideoModal/VideoModal';
 import * as S from './InProgress.style';
 
 interface IncidentCardProps {
@@ -9,40 +10,52 @@ interface IncidentCardProps {
   location: string;
   dateTime: string;
   police: string;
-  // videoUrl: string;
+  videoUrl: string;
 }
 
-const IncidentCard = ({ title, location, dateTime, police }: IncidentCardProps) => {
+const IncidentCard = ({ title, location, dateTime, police, videoUrl }: IncidentCardProps) => {
   const [clickResolve, setClickResolve] = useState(false);
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const openModal = () => {
+    setIsModalOpen(true);
+  };
+
+  const closeModal = () => {
+    setIsModalOpen(false);
+  };
 
   return (
-    <S.FlipCard>
-      <S.CardInner className={clickResolve ? 'is-resolved' : ''}>
-        <S.CardFront>
-          <S.CardHeader>
-            <S.CardTitle>{title} 감지</S.CardTitle>
-            <S.VideoButton>
-              <IoMdCamera className="icon" />
-              영상 확인
-            </S.VideoButton>
-          </S.CardHeader>
-          <S.Line />
-          <S.CardBody>
-            <S.LocationTitle>발생 위치</S.LocationTitle>
-            <S.LocationAddress>{location}</S.LocationAddress>
-            <S.ResolveButton onClick={() => setClickResolve(true)}>사건 해결</S.ResolveButton>
-          </S.CardBody>
-          <S.CardFooter>
-            <span>{dateTime}</span>
-            <span>{police}</span>
-          </S.CardFooter>
-        </S.CardFront>
+    <>
+      <S.FlipCard>
+        <S.CardInner className={clickResolve ? 'is-resolved' : ''}>
+          <S.CardFront>
+            <S.CardHeader>
+              <S.CardTitle>{title} 감지</S.CardTitle>
+              <S.VideoButton onClick={openModal}>
+                <IoMdCamera className="icon" />
+                영상 확인
+              </S.VideoButton>
+            </S.CardHeader>
+            <S.Line />
+            <S.CardBody>
+              <S.LocationTitle>발생 위치</S.LocationTitle>
+              <S.LocationAddress>{location}</S.LocationAddress>
+              <S.ResolveButton onClick={() => setClickResolve(true)}>사건 해결</S.ResolveButton>
+            </S.CardBody>
+            <S.CardFooter>
+              <span>{dateTime}</span>
+              <span>{police}</span>
+            </S.CardFooter>
+          </S.CardFront>
 
-        <S.CardBack>
-          <FeedbackCard onClose={() => setClickResolve(false)} />
-        </S.CardBack>
-      </S.CardInner>
-    </S.FlipCard>
+          <S.CardBack>
+            <FeedbackCard onClose={() => setClickResolve(false)} />
+          </S.CardBack>
+        </S.CardInner>
+      </S.FlipCard>
+      <VideoModal isOpen={isModalOpen} onClose={closeModal} videoUrl={videoUrl} />
+    </>
   );
 };
