+++ frontend/src/mocks/InProgressData.tsx
@@ -4,6 +4,7 @@ export interface Incident {
   location: string;
   dateTime: string;
   police: string;
+  videoUrl: string;
 }
 
 export const InProgressData: Incident[] = [
@@ -13,6 +14,7 @@ export const InProgressData: Incident[] = [
     location: '서울특별시 성북구 보국문로 35길 21-54',
     dateTime: '2025.02.28 15:43:36',
     police: '김하나',
+    videoUrl: 'https://github.com/user-attachments/assets/47bdc8a5-68d1-4edd-8d42-6e06272358b9',
   },
   {
     id: 2,
@@ -20,6 +22,7 @@ export const InProgressData: Incident[] = [
     location: '서울특별시 강남구 테헤란로 87길 12',
     dateTime: '2025.02.28 14:20:10',
     police: '박준영',
+    videoUrl: 'https://github.com/user-attachments/assets/47bdc8a5-68d1-4edd-8d42-6e06272358b9',
   },
   {
     id: 3,
@@ -27,6 +30,7 @@ export const InProgressData: Incident[] = [
     location: '서울특별시 마포구 홍익로 5길 9-7',
     dateTime: '2025.02.28 13:55:22',
     police: '이수민',
+    videoUrl: '',
   },
   {
     id: 4,
@@ -34,6 +38,7 @@ export const InProgressData: Incident[] = [
     location: '서울특별시 성북구 보국문로 35길 21-54',
     dateTime: '2025.02.28 15:43:36',
     police: '김하나',
+    videoUrl: '',
   },
   {
     id: 5,
@@ -41,6 +46,7 @@ export const InProgressData: Incident[] = [
     location: '서울특별시 강남구 테헤란로 87길 12',
     dateTime: '2025.02.28 14:20:10',
     police: '박준영',
+    videoUrl: '',
   },
   {
     id: 6,
@@ -48,6 +54,7 @@ export const InProgressData: Incident[] = [
     location: '서울특별시 마포구 홍익로 5길 9-7',
     dateTime: '2025.02.28 13:55:22',
     police: '이수민',
+    videoUrl: '',
   },
   {
     id: 7,
@@ -55,5 +62,6 @@ export const InProgressData: Incident[] = [
     location: '서울특별시 마포구 홍익로 5길 9-7',
     dateTime: '2025.02.28 13:55:22',
     police: '이수민',
+    videoUrl: '',
   },
 ];
